// impower-dev/src/modules/spark-editor/workspace/types/StorageTypes.ts
export namespace Storage {
// ...
  export interface Drive {
// ...
    name?: string;
// ...
  }
// ...
}